ai/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
pdfplumber==0.10.3
google-genai==0.3.0
python-multipart==0.0.6
python-dotenv==1.0.1
